Adjust 画面遷移
(cherry picked from commit 976317006acb80cb3325a377df306ac10b869499)
