+++ Assets/@Training/_Images/1_Play/Background.meta
