+++ Assets/@Training/_Images/1_Play/Character.meta
