+++ Assets/@Training/_Images/1_Play/Effect.meta
