+++ Assets/@Training/_Images/2_Result/Character.meta
