+++ Assets/@Training/_Images/Common/Character.meta
