+++ Assets/@Training/_Programs/Common/Prefabs/Manager.prefab.meta
