+++ Assets/@Training/_Programs/Common/Scripts/GameManager.cs
@@ -1,0 +1,50 @@
+﻿using UnityEngine;
+using UnityEngine.SceneManagement;
+
+/// <summary>
+/// 画面状態管理クラス
+/// </summary>
+public class GameManager : MonoBehaviour
+{
+    static GameManager instance; // シングルトン用インスタンス
+    /// <summary>
+    /// インスタンスを取得する
+    /// </summary>
+    public static GameManager Instance { get => instance; }
+
+    public enum StateScene
+    {
+        Title,  // タイトル画面
+        Play,   // プレイ画面
+        Result, // 結果画面
+    }
+    StateScene stateScene;
+
+    void Start()
+    {
+        if (instance == null) {
+            // 初回のみインスタンス化
+            instance = new GameManager();
+        } else {
+            // 複製禁止
+            Destroy(gameObject);
+        }
+
+        // シーン遷移による破壊の防止
+        DontDestroyOnLoad(gameObject);
+
+        // 状態の初期化
+        stateScene = (StateScene)SceneManager.GetActiveScene().buildIndex;
+    }
+
+    /// <summary>
+    /// 画面遷移を行う
+    /// </summary>
+    public StateScene ChangeScene
+    {
+        set {
+            stateScene = value;
+            SceneManager.LoadSceneAsync((int)stateScene);
+        }
+    }
+}
