--- Assets/Test.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7c68fcf89c309314298245547bf8d8d5
