--- Assets/_Programs/Sample/SingletonSample.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c757d4c12a274cb4fb7ebdf0c8587f3a
